refactor: move add-phase I/O helpers into node-stack-internal as add_steps module
