+++ crates/core-node-internal/src/services/node/add.rs
@@ -15,10 +15,11 @@ use crate::encoding::{
 };
 use crate::names;
 use chrono::Local;
-use config::consts::{NODE_CONFIG_FILE, PEPPY_OUTPUT_DIR, PEPPYGEN_OUTPUT_PATH, PeppyDirs};
+use config::consts::{NODE_CONFIG_FILE, PEPPYGEN_OUTPUT_PATH, PeppyDirs};
 use config::node::{DEFAULT_VARIANT_NAME, NodeConfig, NodeConfigParser, ParsedNodeConfig};
 use futures::FutureExt;
 use git2::build::RepoBuilder;
+use node_stack::add_steps::{copy_node_to_temp_dir, verify_git_hash};
 use node_stack::{InstanceState, NodeStack, WorkingDirGuard, validate_dependency_specs};
 use parking_lot::Mutex as StdMutex;
 use peppylib::messaging::{ServiceRequestContext, TopicPublisher};
@@ -104,161 +105,6 @@ impl GoalHandler for NodeAddGoalHandler {
         )
         .await
     }
-}
-
-/// Directories excluded from node snapshot copies.
-///
-/// These are either regenerated by the daemon (`.peppy`) or by the node's
-/// `build_cmd` (virtual-environment / package directories). Copying virtual
-/// environments breaks Python on macOS because `std::fs::copy` dereferences
-/// symlinks, turning `@executable_path`-relative dylib references into
-/// dangling paths.
-const COPY_EXCLUDED_DIRS_PEPPY: &[&str] = &[".peppy"];
-
-const COPY_EXCLUDED_DIRS_PYTHON: &[&str] = &[
-    ".venv",
-    ".pixi",
-    "__pypackages__",
-    "__pycache__",
-    ".tox",
-    ".nox",
-    ".conda",
-    ".mypy_cache",
-    ".pytest_cache",
-    ".ruff_cache",
-];
-
-const COPY_EXCLUDED_DIRS_RUST: &[&str] = &["target"];
-
-const COPY_EXCLUDED_DIRS_JS_TS: &[&str] = &[
-    "node_modules",
-    ".next",
-    ".nuxt",
-    ".turbo",
-    ".parcel-cache",
-    ".svelte-kit",
-    "bower_components",
-];
-
-const COPY_EXCLUDED_DIRS_C: &[&str] = &["build", "cmake-build-debug", "cmake-build-release"];
-
-const COPY_EXCLUDED_DIRS_CPP: &[&str] = &[
-    "build",
-    "cmake-build-debug",
-    "cmake-build-release",
-    ".cache",
-];
-
-const COPY_EXCLUDED_DIRS_VCS: &[&str] = &[".git"];
-
-const COPY_EXCLUDED_DIRS: &[&[&str]] = &[
-    COPY_EXCLUDED_DIRS_PEPPY,
-    COPY_EXCLUDED_DIRS_PYTHON,
-    COPY_EXCLUDED_DIRS_RUST,
-    COPY_EXCLUDED_DIRS_JS_TS,
-    COPY_EXCLUDED_DIRS_C,
-    COPY_EXCLUDED_DIRS_CPP,
-    COPY_EXCLUDED_DIRS_VCS,
-];
-
-fn is_excluded_dir(name: &std::ffi::OsStr) -> bool {
-    COPY_EXCLUDED_DIRS
-        .iter()
-        .any(|list| list.iter().any(|&excluded| name == excluded))
-}
-
-/// Recursively copies `src` into `dest`, skipping excluded directories.
-///
-/// Returns the names of top-level excluded directories that were present
-/// directly under `src`. Nested skips are not recorded.
-fn copy_dir_recursive(src: &Path, dest: &Path) -> Result<Vec<String>> {
-    std::fs::create_dir_all(dest)?;
-
-    let mut excluded = Vec::new();
-
-    for entry in std::fs::read_dir(src)? {
-        let entry = entry?;
-        let file_name = entry.file_name();
-
-        if is_excluded_dir(&file_name) {
-            if let Some(name) = file_name.to_str() {
-                excluded.push(name.to_owned());
-            }
-            continue;
-        }
-
-        let src_path = entry.path();
-        let dest_path = dest.join(file_name);
-        let ft = entry.file_type()?;
-
-        if ft.is_symlink() {
-            // Skip symlinks to avoid infinite recursion with circular links
-            continue;
-        } else if ft.is_dir() {
-            copy_dir_recursive(&src_path, &dest_path)?;
-        } else {
-            std::fs::copy(&src_path, &dest_path)?;
-        }
-    }
-    Ok(excluded)
-}
-
-/// Copies a node folder to a temporary working directory.
-///
-/// Returns the path to the temporary directory and the list of top-level
-/// directories that were excluded from the copy. The caller is responsible
-/// for cleaning up the temporary directory.
-fn copy_node_to_temp_dir(from_dir: &Path, tmp_root: &Path) -> Result<(PathBuf, Vec<String>)> {
-    std::fs::create_dir_all(tmp_root)?;
-    let temp_dir = tempfile::TempDir::new_in(tmp_root)?.keep();
-
-    debug!(
-        "Copying node folder from {} to temp dir {}",
-        from_dir.display(),
-        temp_dir.display()
-    );
-
-    let mut excluded = copy_dir_recursive(from_dir, &temp_dir)?;
-    excluded.sort();
-
-    Ok((temp_dir, excluded))
-}
-
-/// Verifies that the node directory is in sync with the currently running daemon
-/// by comparing the stored git hash with the expected one.
-///
-/// Returns `Ok(())` if verification passes, or `Err(error_message)` if it fails.
-fn verify_git_hash(source_path: &Path, expected_git_hash: &str) -> std::result::Result<(), String> {
-    let git_hash_path = source_path.join(PEPPY_OUTPUT_DIR).join("git.hash");
-    let stored_git_hash = std::fs::read_to_string(&git_hash_path).map_err(|e| {
-        format!(
-            "Missing required git hash file at {}: {}. Run `peppy node sync` before `peppy node add`.",
-            git_hash_path.display(),
-            e
-        )
-    })?;
-
-    let expected_git_hash = expected_git_hash.trim();
-    let stored_git_hash = stored_git_hash.trim();
-
-    if stored_git_hash.is_empty() {
-        return Err(format!(
-            "Invalid git hash file at {}: file is empty. Run `peppy node sync` before `peppy node add`.",
-            git_hash_path.display(),
-        ));
-    }
-
-    if stored_git_hash != expected_git_hash {
-        return Err(format!(
-            "git hash mismatch for node directory {} (expected '{}', found '{}' in {}). Run `peppy node sync` before retrying.",
-            source_path.display(),
-            expected_git_hash,
-            stored_git_hash,
-            git_hash_path.display(),
-        ));
-    }
-
-    Ok(())
 }
 
 pub(super) struct CleanupDir(Option<PathBuf>);
