+++ crates/node-stack-internal/src/lib.rs
@@ -9,6 +9,7 @@ pub use error::Error as NodeStackError;
 
 pub use archive::extract_tar_zst;
 pub use build_io::{FeedbackLine, FeedbackStream, OutputReaderHooks};
+pub use node_stack::add_steps;
 pub use node_stack::{
     BuildContext, DependencySpec, EntityHandle, EntitySnapshot, InstanceState, NodeEntity,
     NodeStack, NodeStage, OutputSinks, RestoreTarget, SerializedNodeGraph, StartContext,
