+++ crates/node-stack-internal/src/node_stack.rs
@@ -1,3 +1,4 @@
+pub mod add_steps;
 mod build_steps;
 mod entity;
 mod start_steps;
