+++ crates/node-stack-internal/src/node_stack/add_steps.rs
@@ -1,0 +1,308 @@
+//! Concrete add I/O steps invoked from the add orchestration in
+//! `core-node-internal`. Lives here so add-phase I/O primitives sit next to
+//! the entity state they populate ([`super::entity::NodeStage::Added`],
+//! `pending_working_dir`).
+
+use std::path::{Path, PathBuf};
+
+use config::consts::PEPPY_OUTPUT_DIR;
+use tracing::debug;
+
+/// Directories excluded from node snapshot copies.
+///
+/// These are either regenerated by the daemon (`.peppy`) or by the node's
+/// `build_cmd` (virtual-environment / package directories). Copying virtual
+/// environments breaks Python on macOS because `std::fs::copy` dereferences
+/// symlinks, turning `@executable_path`-relative dylib references into
+/// dangling paths.
+const COPY_EXCLUDED_DIRS_PEPPY: &[&str] = &[".peppy"];
+
+const COPY_EXCLUDED_DIRS_PYTHON: &[&str] = &[
+    ".venv",
+    ".pixi",
+    "__pypackages__",
+    "__pycache__",
+    ".tox",
+    ".nox",
+    ".conda",
+    ".mypy_cache",
+    ".pytest_cache",
+    ".ruff_cache",
+];
+
+const COPY_EXCLUDED_DIRS_RUST: &[&str] = &["target"];
+
+const COPY_EXCLUDED_DIRS_JS_TS: &[&str] = &[
+    "node_modules",
+    ".next",
+    ".nuxt",
+    ".turbo",
+    ".parcel-cache",
+    ".svelte-kit",
+    "bower_components",
+];
+
+const COPY_EXCLUDED_DIRS_C: &[&str] = &["build", "cmake-build-debug", "cmake-build-release"];
+
+const COPY_EXCLUDED_DIRS_CPP: &[&str] = &[
+    "build",
+    "cmake-build-debug",
+    "cmake-build-release",
+    ".cache",
+];
+
+const COPY_EXCLUDED_DIRS_VCS: &[&str] = &[".git"];
+
+const COPY_EXCLUDED_DIRS: &[&[&str]] = &[
+    COPY_EXCLUDED_DIRS_PEPPY,
+    COPY_EXCLUDED_DIRS_PYTHON,
+    COPY_EXCLUDED_DIRS_RUST,
+    COPY_EXCLUDED_DIRS_JS_TS,
+    COPY_EXCLUDED_DIRS_C,
+    COPY_EXCLUDED_DIRS_CPP,
+    COPY_EXCLUDED_DIRS_VCS,
+];
+
+fn is_excluded_dir(name: &std::ffi::OsStr) -> bool {
+    COPY_EXCLUDED_DIRS
+        .iter()
+        .any(|list| list.iter().any(|&excluded| name == excluded))
+}
+
+/// Recursively copies `src` into `dest`, skipping excluded directories.
+///
+/// Returns the names of top-level excluded directories that were present
+/// directly under `src`. Nested skips are not recorded.
+pub fn copy_dir_recursive(src: &Path, dest: &Path) -> std::io::Result<Vec<String>> {
+    std::fs::create_dir_all(dest)?;
+
+    let mut excluded = Vec::new();
+
+    for entry in std::fs::read_dir(src)? {
+        let entry = entry?;
+        let file_name = entry.file_name();
+
+        if is_excluded_dir(&file_name) {
+            if let Some(name) = file_name.to_str() {
+                excluded.push(name.to_owned());
+            }
+            continue;
+        }
+
+        let src_path = entry.path();
+        let dest_path = dest.join(file_name);
+        let ft = entry.file_type()?;
+
+        if ft.is_symlink() {
+            // Skip symlinks to avoid infinite recursion with circular links
+            continue;
+        } else if ft.is_dir() {
+            copy_dir_recursive(&src_path, &dest_path)?;
+        } else {
+            std::fs::copy(&src_path, &dest_path)?;
+        }
+    }
+    Ok(excluded)
+}
+
+/// Copies a node folder to a temporary working directory.
+///
+/// Returns the path to the temporary directory and the list of top-level
+/// directories that were excluded from the copy. The caller is responsible
+/// for cleaning up the temporary directory.
+pub fn copy_node_to_temp_dir(
+    from_dir: &Path,
+    tmp_root: &Path,
+) -> std::io::Result<(PathBuf, Vec<String>)> {
+    std::fs::create_dir_all(tmp_root)?;
+    let temp_dir = tempfile::TempDir::new_in(tmp_root)?.keep();
+
+    debug!(
+        "Copying node folder from {} to temp dir {}",
+        from_dir.display(),
+        temp_dir.display()
+    );
+
+    let mut excluded = copy_dir_recursive(from_dir, &temp_dir)?;
+    excluded.sort();
+
+    Ok((temp_dir, excluded))
+}
+
+/// Verifies that the node directory is in sync with the currently running daemon
+/// by comparing the stored git hash with the expected one.
+///
+/// Returns `Ok(())` if verification passes, or `Err(error_message)` if it fails.
+pub fn verify_git_hash(
+    source_path: &Path,
+    expected_git_hash: &str,
+) -> std::result::Result<(), String> {
+    let git_hash_path = source_path.join(PEPPY_OUTPUT_DIR).join("git.hash");
+    let stored_git_hash = std::fs::read_to_string(&git_hash_path).map_err(|e| {
+        format!(
+            "Missing required git hash file at {}: {}. Run `peppy node sync` before `peppy node add`.",
+            git_hash_path.display(),
+            e
+        )
+    })?;
+
+    let expected_git_hash = expected_git_hash.trim();
+    let stored_git_hash = stored_git_hash.trim();
+
+    if stored_git_hash.is_empty() {
+        return Err(format!(
+            "Invalid git hash file at {}: file is empty. Run `peppy node sync` before `peppy node add`.",
+            git_hash_path.display(),
+        ));
+    }
+
+    if stored_git_hash != expected_git_hash {
+        return Err(format!(
+            "git hash mismatch for node directory {} (expected '{}', found '{}' in {}). Run `peppy node sync` before retrying.",
+            source_path.display(),
+            expected_git_hash,
+            stored_git_hash,
+            git_hash_path.display(),
+        ));
+    }
+
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::fs;
+
+    #[test]
+    fn is_excluded_dir_rejects_known_dirs() {
+        for name in [
+            ".peppy",
+            ".venv",
+            "node_modules",
+            "target",
+            ".git",
+            "build",
+            "__pycache__",
+        ] {
+            assert!(
+                is_excluded_dir(std::ffi::OsStr::new(name)),
+                "expected {:?} to be excluded",
+                name
+            );
+        }
+    }
+
+    #[test]
+    fn is_excluded_dir_accepts_normal_dirs() {
+        for name in ["src", "lib", "data", "scripts", "peppy", "my_module"] {
+            assert!(
+                !is_excluded_dir(std::ffi::OsStr::new(name)),
+                "expected {:?} to be allowed",
+                name
+            );
+        }
+    }
+
+    #[test]
+    fn copy_dir_recursive_excludes_and_copies() {
+        let src = tempfile::tempdir().expect("tempdir");
+        let dest = tempfile::tempdir().expect("tempdir");
+
+        // Create included and excluded dirs/files
+        fs::create_dir(src.path().join("src")).unwrap();
+        fs::write(src.path().join("src/main.rs"), "fn main() {}").unwrap();
+        fs::write(src.path().join("peppy.json5"), "{}").unwrap();
+        fs::create_dir(src.path().join("target")).unwrap();
+        fs::write(src.path().join("target/debug"), "binary").unwrap();
+        fs::create_dir(src.path().join(".git")).unwrap();
+        fs::write(src.path().join(".git/HEAD"), "ref").unwrap();
+
+        let excluded = copy_dir_recursive(src.path(), dest.path()).expect("copy should succeed");
+
+        assert!(excluded.contains(&"target".to_string()));
+        assert!(excluded.contains(&".git".to_string()));
+        assert!(dest.path().join("src/main.rs").exists());
+        assert!(dest.path().join("peppy.json5").exists());
+        assert!(!dest.path().join("target").exists());
+        assert!(!dest.path().join(".git").exists());
+    }
+
+    #[test]
+    fn copy_dir_recursive_skips_symlinks() {
+        let src = tempfile::tempdir().expect("tempdir");
+        let dest = tempfile::tempdir().expect("tempdir");
+
+        fs::write(src.path().join("real.txt"), "content").unwrap();
+        #[cfg(unix)]
+        {
+            std::os::unix::fs::symlink(src.path().join("real.txt"), src.path().join("link.txt"))
+                .unwrap();
+        }
+
+        copy_dir_recursive(src.path(), dest.path()).expect("copy should succeed");
+
+        assert!(dest.path().join("real.txt").exists());
+        #[cfg(unix)]
+        assert!(!dest.path().join("link.txt").exists());
+    }
+
+    #[test]
+    fn copy_node_to_temp_dir_creates_temp_and_copies() {
+        let src = tempfile::tempdir().expect("tempdir");
+        let tmp_root = tempfile::tempdir().expect("tempdir");
+
+        fs::write(src.path().join("peppy.json5"), "{}").unwrap();
+        fs::create_dir(src.path().join(".venv")).unwrap();
+
+        let (working_dir, excluded) =
+            copy_node_to_temp_dir(src.path(), tmp_root.path()).expect("copy should succeed");
+
+        assert!(working_dir.join("peppy.json5").exists());
+        assert!(!working_dir.join(".venv").exists());
+        assert!(excluded.contains(&".venv".to_string()));
+
+        // Clean up the kept tempdir
+        fs::remove_dir_all(&working_dir).ok();
+    }
+
+    #[test]
+    fn verify_git_hash_succeeds_on_match() {
+        let dir = tempfile::tempdir().expect("tempdir");
+        let peppy_dir = dir.path().join(PEPPY_OUTPUT_DIR);
+        fs::create_dir_all(&peppy_dir).unwrap();
+        fs::write(peppy_dir.join("git.hash"), "abc123\n").unwrap();
+
+        assert!(verify_git_hash(dir.path(), "abc123").is_ok());
+    }
+
+    #[test]
+    fn verify_git_hash_fails_on_mismatch() {
+        let dir = tempfile::tempdir().expect("tempdir");
+        let peppy_dir = dir.path().join(PEPPY_OUTPUT_DIR);
+        fs::create_dir_all(&peppy_dir).unwrap();
+        fs::write(peppy_dir.join("git.hash"), "abc123").unwrap();
+
+        let err = verify_git_hash(dir.path(), "def456").unwrap_err();
+        assert!(err.contains("git hash mismatch"), "got: {err}");
+    }
+
+    #[test]
+    fn verify_git_hash_fails_on_empty() {
+        let dir = tempfile::tempdir().expect("tempdir");
+        let peppy_dir = dir.path().join(PEPPY_OUTPUT_DIR);
+        fs::create_dir_all(&peppy_dir).unwrap();
+        fs::write(peppy_dir.join("git.hash"), "  \n").unwrap();
+
+        let err = verify_git_hash(dir.path(), "abc123").unwrap_err();
+        assert!(err.contains("file is empty"), "got: {err}");
+    }
+
+    #[test]
+    fn verify_git_hash_fails_on_missing_file() {
+        let dir = tempfile::tempdir().expect("tempdir");
+
+        let err = verify_git_hash(dir.path(), "abc123").unwrap_err();
+        assert!(err.contains("Missing required git hash file"), "got: {err}");
+    }
+}
